Update Sales_GUI.py
In this update, I have added a function that will allow the users to clear the entries, they can clear all entries one time, one click. I will be working on the next update, which is to clear all fields in the right frame after a receipt bill has been printed and users have clicked on the clear button.